FRAMEWORK CHANGE: HUGO
I thought I would be using Jekyll for static site generation but
I have struggled to use Jekyll so I have switched to Hugo for a
more performant and easier use-case. Hugo has been easy to setup
with a few caveats. The module functionality was a bit funky but
I somehow go through it. Now it's working with Hugo with Terminal
theme, which looks amazing in my opinion.